Refactor: login state, sideeffect를 contract 패키지로 이동
